입력값을 포인트로 변환하기 위해 구조체 분리
